refactor + unit test
