+++ dbt/adapters/bigquery/column.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Optional, List, TypeVar, Iterable, Type, Any
+from typing import Optional, List, TypeVar, Iterable, Type, Any, Dict, Union
 
 from dbt.adapters.base.column import Column
 
@@ -126,3 +126,122 @@ class BigQueryColumn(Column):
             kwargs = {"fields": fields}
 
         return SchemaField(self.name, self.dtype, self.mode, **kwargs)  # type: ignore[arg-type]
+
+
+def get_nested_column_data_types(
+    columns: Dict[str, Dict[str, Any]],
+    constraints: Optional[Dict[str, str]] = None,
+) -> Dict[str, Dict[str, str]]:
+    """
+    columns:
+        * Dictionary where keys are of flat columns names and values are dictionary of column attributes
+        * column names with "." indicate a nested column within a STRUCT type
+        * e.g. {"a": {"name": "a", "data_type": "string", ...}}
+    constraints:
+        * Dictionary where keys are flat column names and values are rendered constraints for the column
+        * If provided, rendered column is included in returned "data_type" values.
+    returns:
+        * Dictionary where keys are root column names and values are corresponding nested data_type values.
+        * Fields other than "name" and "data_type" are not preserved in the return value.
+
+    Example:
+    columns: {
+        "a": {"name": "a", "data_type": "string", "description": ...},
+        "b.nested": {"name": "b.nested", "data_type": "string"},
+        "b.nested2": {"name": "b.nested2", "data_type": "string"}
+        }
+
+    returns: {
+        "a": {"name": "a", "data_type": "string"},
+        "b": {"name": "b": "data_type": "struct<nested string, nested2 string>}
+    }
+    """
+    constraints = constraints or {}
+
+    nested_column_data_types: Dict[str, Union[str, Dict]] = {}
+    for column in columns.values():
+        _update_nested_column_data_types(
+            column["name"],
+            column["data_type"],
+            constraints.get(column["name"]),
+            nested_column_data_types,
+        )
+
+    formatted_nested_column_data_types: Dict[str, Dict[str, str]] = {}
+    for column_name, unformatted_column_type in nested_column_data_types.items():
+        formatted_nested_column_data_types[column_name] = {
+            "name": column_name,
+            "data_type": _format_nested_data_type(unformatted_column_type),
+        }
+
+    return formatted_nested_column_data_types
+
+
+def _update_nested_column_data_types(
+    column_name: str,
+    column_data_type: str,
+    column_rendered_constraint: Optional[str],
+    nested_column_data_types: Dict[str, Union[str, Dict]],
+) -> None:
+    """
+    Recursively update nested_column_data_types given a column_name, column_data_type, and optional column_rendered_constraint.
+
+    Examples:
+    >>> nested_column_data_types = {}
+    >>> BigQueryAdapter._update_nested_column_data_types("a", "string", "not_null", nested_column_data_types)
+    >>> nested_column_data_types
+    {"a": "string not null"}
+    >>> BigQueryAdapter._update_nested_column_data_types("b.c", "string", "not_null", nested_column_data_types)
+    >>> nested_column_data_types
+    {"a": "string not null", "b": {"c": "string not null"}}
+    >>> BigQueryAdapter._update_nested_column_data_types("b.d", "string", None, nested_column_data_types)
+    >>> nested_column_data_types
+    {"a": "string not null", "b": {"c": "string not null", "d": "string"}}
+    """
+    column_name_parts = column_name.split(".")
+    root_column_name = column_name_parts[0]
+
+    if len(column_name_parts) == 1:
+        # Base case: column is not nested - store its data_type concatenated with constraint if provided.
+        nested_column_data_types[root_column_name] = (
+            column_data_type
+            if column_rendered_constraint is None
+            else f"{column_data_type} {column_rendered_constraint}"
+        )
+    else:
+        # Initialize nested dictionary
+        if root_column_name not in nested_column_data_types:
+            nested_column_data_types[root_column_name] = {}
+
+        # Recursively process rest of remaining column name
+        remaining_column_name = ".".join(column_name_parts[1:])
+        remaining_column_data_types = nested_column_data_types[root_column_name]
+        assert isinstance(remaining_column_data_types, dict)  # keeping mypy happy
+        _update_nested_column_data_types(
+            remaining_column_name,
+            column_data_type,
+            column_rendered_constraint,
+            remaining_column_data_types,
+        )
+
+
+def _format_nested_data_type(unformatted_nested_data_type: Union[str, Dict[str, Any]]) -> str:
+    """
+    Recursively format a (STRUCT) data type given an arbitrarily nested data type structure.
+
+    Examples:
+    >>> BigQueryAdapter._format_nested_data_type("string")
+    'string'
+    >>> BigQueryAdapter._format_nested_data_type({'c': 'string not_null', 'd': 'string'})
+    'struct<c string not_null, d string>'
+    >>> BigQueryAdapter._format_nested_data_type({'c': 'string not_null', 'd': {'e': 'string'}})
+    'struct<c string not_null, d struct<e string>>'
+    """
+    if isinstance(unformatted_nested_data_type, str):
+        return unformatted_nested_data_type
+    else:
+        formatted_nested_types = [
+            f"{column_name} {_format_nested_data_type(column_type)}"
+            for column_name, column_type in unformatted_nested_data_type.items()
+        ]
+        return f"""struct<{", ".join(formatted_nested_types)}>"""
